+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.1.1 (2025-04-07)
+* Restructured navigation handler
+* Added link target handling
+* Added button and form target handling
+* Always load navigation handler
+* Implemented view transitions
+
 ## v0.1.0 (2025-04-04)
 * Initial component implementation
 * Built initial Page island structure